@lostluck 2022-01-20
I think a small change to also having a io buffer to the more complicated encodings, will also improve hashing for custom coders and the like. I'll try it, but send a different PR for it.